Cleaned map_data file
This file converts 4-digit county FIPS codes to 5-digit codes so Tableau can read them all.
NOTE: When opened in Excel, Excel has a default message that asks if you want to convert some numbers by removing the leading zero. Do not convert as that will remove the added numeral creating the 5-digit code.
Also, use the second column, named New_FIPS, which contains the 4- and 5-digit codes in one column.
Finally, the two FIPS columns will show an error tag that the two columns are formatted as text, not numbers. Tableau wanted those as text, not numbers, so ignore that error prompt.